requirement.txt
django==1.10.5
djangorestframework==3.9.2
django-summernote==0.8.11.4